backend/requirements.txt
babel==2.17.0
blinker==1.9.0
click==8.2.1
docxcompose==1.4.0
docxtpl==0.20.1
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
python-docx==1.2.0
six==1.17.0
typing_extensions==4.14.1
Werkzeug==3.1.3
